backend/requirements.txt
secure-smtpd
Flask
Flask-Cors
Flask-MySQL
Flask-Redis
PyMySQL
pandas
numpy
scipy
dateparser
xlrd
sqlalchemy
flask-restful
jsonschema==2.6.0
flasgger
flask-sqlalchemy
schedule
rsa
python-gnupg
xlwt
openpyxl
sqlacodegen
Flask-Migrate
py2neo
XlsxWriter
requests
kafka-python
networkx
rocketmq
uwsgi
pycurl